improve looking game

+++ Functions/FisrtGameBet.py
@@ -13,7 +13,7 @@ from Functions.ValidationDuParis import ValidationDuParis
 def FirstGameBet(driver):
     config.log('PREPARATTION PREMIER PARIS', 'title', False, 1)
     config.newset = int(config.set_actuel) + 1
-    config.looking_game = config.jeu_actuel
+    config.looking_game = int(config.jeu_actuel)
     bet_40a = False
     tentative = 0
     nextBet = False
@@ -23,19 +23,19 @@ def FirstGameBet(driver):
             if config.score_actuel != "0:0" and config.score_actuel != "0:15" and config.score_actuel != "15:0" and config.score_actuel != "15:15":
                 config.log(f'       score : {config.score_actuel} ...1er jeu passé !', 'warning', True)
                 nextBet = True
-                config.looking_game = float(config.jeu_actuel) + 1
+                config.looking_game = int(config.jeu_actuel) + 1
         elif config.scriptType == '15A' or config.scriptType == '400' or config.scriptType == '030' or config.scriptType == '300' or config.scriptType == '6P' or config.scriptType == '5P' or config.scriptType == '4P':
             if config.score_actuel != "0:0":
                 config.log(f'       score : {config.score_actuel} ...1er jeu passé !', 'warning', True)
                 nextBet = True
-                config.looking_game = float(config.jeu_actuel) + 1
+                config.looking_game = int(config.jeu_actuel) + 1
         elif config.scriptType == '40A':
             if config.score_actuel == "40:40" or config.score_actuel == "A:40" or config.score_actuel == "40:A":
                 config.log(f'       score : {config.score_actuel} ...1er jeu passé !', 'warning', True)
                 nextBet = True
-                config.looking_game = float(config.jeu_actuel) + 1
+                config.looking_game = int(config.jeu_actuel) + 1
         else:
-            config.looking_game = float(config.jeu_actuel)
+            config.looking_game = int(config.jeu_actuel)
 
         config.log(f'Affichage de la liste des paris', 'info', True, 2)
