+++ Functions/GetAndPlaceBet.py
@@ -15,17 +15,17 @@ def GetAndPlaceBet(driver):
         if config.scriptType == '30A':
             if config.score_actuel != "0:0" and config.score_actuel != "0:15" and config.score_actuel != "15:0" and config.score_actuel != "15:15":
                 config.log(f'       score : {config.score_actuel} ...1er jeu passé !', 'warning', True)
-                config.looking_game = float(config.jeu_actuel) + 1
+                config.looking_game = int(config.jeu_actuel) + 1
         elif config.scriptType == '15A' or config.scriptType == '400' or config.scriptType == '030' or config.scriptType == '300' or config.scriptType == '6P' or config.scriptType == '5P' or config.scriptType == '4P':
             if config.score_actuel != "0:0":
                 config.log(f'       score : {config.score_actuel} ...1er jeu passé !', 'warning', True)
-                config.looking_game = float(config.jeu_actuel) + 1
+                config.looking_game = int(config.jeu_actuel) + 1
         elif config.scriptType == '40A':
             if config.score_actuel == "40:40" or config.score_actuel == "A:40" or config.score_actuel == "40:A":
                 config.log(f'       score : {config.score_actuel} ...1er jeu passé !', 'warning', True)
-                config.looking_game = float(config.jeu_actuel) + 1
+                config.looking_game = int(config.jeu_actuel) + 1
         else:
-            config.looking_game = float(config.jeu_actuel)
+            config.looking_game = int(config.jeu_actuel)
         config.log(f'jeu recherhcé : {config.looking_game}', 'info', True)
         # Affichage de la liste des paris
         config.log('Affichage de la liste des paris', config.newmatch)
